+++ lib/src/simple_card/simple_card_component.dart
@@ -47,6 +47,11 @@ class EssentialSimpleCardComponent implements AfterContentInit {
 
   @Output()
   Stream<bool> get dataRequest => _dataRequest.stream;
+
+  final _linkClick = StreamController<html.Event>();
+
+  @Output()
+  Stream<html.Event> get linkClick => _linkClick.stream;
   
   @Input('data')
   SimpleCardRender data;
@@ -71,8 +76,14 @@ class EssentialSimpleCardComponent implements AfterContentInit {
   /*
    * Essas ações estarão disponíveis para o desenvolvedor
    */
+
+  @Input()
   bool showMinimizeBtn = true;
+  
+  @Input()
   bool showRefreshBtn = true;
+  
+  @Input()
   bool showCloseBtn = true;
 
   /*
@@ -119,22 +130,27 @@ class EssentialSimpleCardComponent implements AfterContentInit {
     return displayMinimize == true;
   }
 
-  toggleClose() {
+  void toggleClose() {
     hiddenClose = !hiddenClose;
   }
 
   bool isLoading = true;
 
-  reload() {
+  void reload() {
     _dataRequest.add(true);
   }
 
-  goToUrl() {
-    print(this.model.linkUrl);
+  void goToUrl() {
+    html.window.
+    print();
   }
 
   bool hasButton() {
     return model.templateLink == TemplateLink.BUTTON;
+  }
+
+  handleClick(event) {
+    _linkClick.add(event);
   }
 
 }